Fix casting to size_t instead of uint32 for a pointer.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@69600 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/Threads/Locks.h
@@ -185,7 +185,7 @@ public:
   TFatLock* changeToFatlock(Owner* O) {
     if (!(lock & FatMask)) {
       TFatLock* obj = TFatLock::allocate(O);
-      uint32 val = (((uint32) obj) >> 1) | FatMask;
+      size_t val = (((size_t) obj) >> 1) | FatMask;
       uint32 count = lock & 0xFF;
       obj->acquireAll(count + 1);
       lock = val;
